# api/src/static/api/api.yml
openapi: 3.0.0
info:
  title: User Service Database API
  description: 'Api rest to register users in a database for applications. Go to [Home page](http://localhost:5000/) for more information.'
  version: 1.0.0
  termsOfService: 'http://localhost:5000'
  contact:
    email: rodrjuniorsantos@gmail.com
servers:
  - url: 'http://localhost:5000'
    description: 'Primary server, e.g. Main (production) server'
components:
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: api-token
paths:
...
  /v1/api/addresss/:
    post:
      summary: Create Adress Function.
      security:
        - APIKeyHeader: []
      tags:
        - Address
      requestBody:
        description: Adress Functions
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - street
                - city
                - zipCode
                - state
                - number
              properties:
                street:
                  type: string
                city:
                  type: integer
                zipCode:
                  type: integer
                state:
                  type: string
                number:
                  type: integer
                deleted_app:
                  type: integer
                created_app:
                  type: integer
                modified_app:
                  type: integer
      responses:
        '200':
          description: Adress successfully registered
        '400':
          description: Error in request
        '401':
          description: Entity Error
        '402':
          description: Missing data
    get:
      summary: Get all address Function
      security:
        - APIKeyHeader: []
      tags:
        - Address
      
      responses:
        '200':
          description: Returns all address
        '400':
          description: User not found
        '401':
          description: Permission denied
  /v1/api/addresss/{address_id}:
    get:
      summary: Gets a address by ID.
      security:
        - APIKeyHeader: []
      tags:
        - Address
      parameters:
        - in: path
          name: address_id
          required: true
          schema:
            type: integer
            minimum: 1
            description: The address ID.

      responses:
        '200':
          description: Data this address successfully
        '400':
          description: user not found
        '401':
          description: Permission denied
    put:
      summary: Update A Address.
      security:
        - APIKeyHeader: []
      tags:
        - Address
      parameters:
        - in: path
          name: address_id
          required: true
          schema:
            type: integer
            minimum: 1
            description: The Address ID.
      requestBody:
        description: Adress Functions
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - street
                - city
                - zipCode
                - state
                - number
              properties:
                street:
                  type: string
                city:
                  type: integer
                zipCode:
                  type: integer
                state:
                  type: string
                number:
                  type: integer
                deleted_app:
                  type: integer
                created_app:
                  type: integer
                modified_app:
                  type: integer
      responses:
        '200':
          description: Address successfully update
        '400':
          description: Error in request
        '401':
          description: Address not found
        '402':
          description: Permission denied
        '403':
          description: Missing data
    delete:
      summary: Delete a address by ID.
      security:
        - APIKeyHeader: []
      tags:
        - Address
      parameters:
        - in: path
          name: address_id
          required: true
          schema:
            type: integer
            minimum: 1
            description: The address ID.
      responses:
        '200':
          description: Address successfully deleted
        '400':
          description: Address not found
        '401':
          description: Permission denied
  /v1/api/apps/:
    post:
      summary: Create App Function.
      security:
        - APIKeyHeader: []
      tags:
        - App
      requestBody:
        description: App Functions
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - description
              properties:
                name:
                  type: string
                description:
                  type: string
                deleted_app:
                  type: integer
                created_app:
                  type: integer
                modified_app:
                  type: integer
      responses:
        '200':
          description: App successfully registered
        '400':
          description: Error in request
        '401':
          description: Missing data
    get:
      summary: Get all apps Function
      security:
        - APIKeyHeader: []
      tags:
        - App
      
      responses:
        '200':
          description: Returns all apps
        '400':
          description: User not found
        '401':
          description: Permission denied
  /v1/api/apps/{app_id}:
    get:
      summary: Gets a app by ID.
      security:
        - APIKeyHeader: []
      tags:
        - App
      parameters:
        - in: path
          name: app_id
          required: true
          schema:
            type: integer
            minimum: 1
            description: The app ID.

      responses:
        '200':
          description: Data this app successfully
        '400':
          description: user not found
        '401':
          description: Permission denied
    put:
      summary: Update A Profesion.
      security:
        - APIKeyHeader: []
      tags:
        - App
      parameters:
        - in: path
          name: app_id
          required: true
          schema:
            type: integer
            minimum: 1
            description: The app ID.
      requestBody:
        description: App Functions
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - description
              properties:
                name:
                  type: string
                description:
                  type: string
                deleted_app:
                  type: integer
                created_app:
                  type: integer
                modified_app:
                  type: integer
      responses:
        '200':
          description: Profesion successfully update
        '400':
          description: Error in request
        '401':
          description: App not found
        '402':
          description: Permission denied
        '403':
          description: Missing data
    delete:
      summary: Delete a app by ID.
      security:
        - APIKeyHeader: []
      tags:
        - App
      parameters:
        - in: path
          name: app_id
          required: true
          schema:
            type: integer
            minimum: 1
            description: The app ID.
      responses:
        '200':
          description: App successfully deleted
        '400':
          description: App not found
        '401':
          description: Permission denied
...